refactor otel audit to its own file
